refactor: remove obsolete bills lib
